+++ src/widgets/data-request-wizard/data-request-wizard-consumer/data-request-wizard-consumer.component.ts
@@ -1,4 +1,3 @@
-import { CommonModule } from '@angular/common';
 import { Component, computed, effect, Signal, signal, untracked } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { FontAwesomeModule } from '@fortawesome/angular-fontawesome';
@@ -35,7 +34,6 @@ import { DataRequestWizardBaseComponent } from '@/widgets/data-request-wizard/da
   imports: [
     AgridataWizardComponent,
     ButtonComponent,
-    CommonModule,
     DataRequestCompletionComponent,
     DataRequestFormConsumerComponent,
     DataRequestFormContractComponent,